authorization add

+++ controllers/admin.js
@@ -1,7 +1,7 @@
 const Product = require('../models/product');
 
 exports.getProducts = (req, res, next) => {
-  Product.find({}).populate('userId')
+  Product.find({usrId:req.user._id}).populate('userId')
     .then(products => {
       res.render('admin/products', {
         prods: products,
@@ -76,23 +76,21 @@ exports.postEditProduct = (req, res, next) => {
 
   Product.findById(prodId)
     .then(product => {
+      if(product.userId.toString() !==req.user._id.toString()){ return res.redirect('/')}
       product.title = updatedTitle
       product.price = updatedPrice
       product.imageUrl = updatedImageUrl
       product.description = updatedDesc
-      return product.save()
-    })
-    .then(result => {
-      res.redirect('/admin/products');
-    })
+      return product.save().then(result => {
+        res.redirect('/admin/products');
+      })
+    })    
     .catch(err => console.log(err));
 };
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({
-      _id: prodId
-    })
+  Product.deleteOne({_id: prodId, userId:req.user._id })
     .then(() => {
       console.log('Product Deleted');
       res.redirect('/admin/products');
